Fixed windows compilation removing warnings

+++ lib/src/DosQMetaObject.cpp
@@ -30,7 +30,7 @@ QList<QByteArray> createParameterNames(const DOS::SignalDefinition &signal)
 {
     QList<QByteArray> result;
     const auto &parameters = signal.parameters;
-    result.reserve(parameters.size());
+    result.reserve(static_cast<int>(parameters.size()));
     for (size_t i = 0; i < parameters.size(); ++i)
         result << parameters[i].name.toUtf8();
     return result;
@@ -100,17 +100,17 @@ const QMetaObject *BaseDosQMetaObject::metaObject() const
     return m_metaObject;
 }
 
-QMetaMethod BaseDosQMetaObject::signal(const QString &/*signalName*/) const
+QMetaMethod BaseDosQMetaObject::signal(const QString &) const
 {
     return QMetaMethod();
 }
 
-QMetaMethod BaseDosQMetaObject::readSlot(const char */*propertyName*/) const
+QMetaMethod BaseDosQMetaObject::readSlot(const char *) const
 {
     return QMetaMethod();
 }
 
-QMetaMethod BaseDosQMetaObject::writeSlot(const char */*propertyName*/) const
+QMetaMethod BaseDosQMetaObject::writeSlot(const char *) const
 {
     return QMetaMethod();
 }
